requirements.txt
requests
pyOpenSSL
python-gnupg